DEV-797 Parse image in the same way as figure (#194)
