+++ packages/gitbook-docs/src/renderContent.ts
@@ -4,7 +4,7 @@ import { ReactShape } from './ReactShape';
 import { SwaggerProps } from './markdoc/schema/swagger';
 import { LinkProps } from './markdoc/schema/link';
 import { HintProps } from './markdoc/schema/hint';
-import { FigureProps } from './markdoc/schema/figure';
+import { ImageProps } from './markdoc/schema/image';
 import { ParagraphProps } from './markdoc/schema/paragraph';
 import { HeadingProps } from './markdoc/schema/heading';
 import { ListProps } from './markdoc/schema/list';
@@ -23,7 +23,7 @@ export type RenderingComponents<A> = {
   // readonly Quote: (props: QuoteProps<A>) => A;
   readonly CodeBlock: (props: CodeBlockProps<A>) => A;
   // readonly File: (props: FileProps<A>) => A;
-  readonly Figure: (props: FigureProps<A>) => A;
+  readonly Image: (props: ImageProps) => A;
   // readonly Embed: (props: EmbedProps<A>) => A;
   // readonly Table: (props: TableProps<A>) => A;
   // readonly Cards: (props: CardsProps<A>) => A;
